move gui command classes
