+++ src/python/cuemol_gui/commands/gui_command_base.py
@@ -1,0 +1,9 @@
+class GUICommandBase(object):
+    def __init__(self):
+        pass
+
+    def run(self, widget, undo_txn):
+        raise NotImplementedError()
+
+    def get_name(self):
+        raise NotImplementedError()
