+++ src/python/cuemol_gui/commands/gui_command_manager.py
@@ -1,17 +1,6 @@
 import copy
 
 
-class GUICommandBase(object):
-    def __init__(self):
-        pass
-
-    def run(self, widget, undo_txn):
-        raise NotImplementedError()
-
-    def get_name(self):
-        raise NotImplementedError()
-
-
 class GUICommandManager(object):
     def __init__(self):
         self._data = {}
